Prettify, lint, and fix warnings

+++ contracts/V1ColorCoin.sol
@@ -4,34 +4,30 @@ pragma solidity ^0.8.9;
 import "@openzeppelin/contracts/token/ERC721/ERC721.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 import "@openzeppelin/contracts/utils/Counters.sol";
-import "@openzeppelin/contracts/utils/Strings.sol"; 
+import "@openzeppelin/contracts/utils/Strings.sol";
 import "@openzeppelin/contracts/utils/Base64.sol";
-
 
 contract ColorCoin is ERC721, Ownable {
     using Counters for Counters.Counter;
 
     Counters.Counter private _tokenIdCounter;
-    mapping(uint => uint) public huesById;
+    mapping(uint256 => uint256) public huesById;
 
     /**
-     * Mints 100 NFTs, a list of recipients will recieve the first NFTs, and the rest will go to the creator's wallet. 
+     * Mints 100 NFTs, a list of recipients will recieve the first NFTs, and the rest will go to the creator's wallet.
      **/
     constructor(address[] memory recipients) ERC721("ColorCoin", "CLC") {
-
-        for (uint i = 0; i < 5; i++) {
-
-            if (i < recipients.length){
+        for (uint256 i = 0; i < 5; i++) {
+            if (i < recipients.length) {
                 safeMint(recipients[i]);
-            }
-            else {
+            } else {
                 safeMint(msg.sender);
             }
         }
-
     }
+
     /**
-     * Mints a token with an incremented ID. 
+     * Mints a token with an incremented ID.
      **/
     function safeMint(address to) public onlyOwner {
         uint256 tokenId = _tokenIdCounter.current();
@@ -41,69 +37,81 @@ contract ColorCoin is ERC721, Ownable {
     }
 
     /**
-     * Generates a svg, using a data URI in order to provide a thumbnail to the image generated for the NFT. 
-     **/    
-    function generateImage (uint hue, uint saturation, uint lightness) public pure returns (string memory) {
+     * Generates a svg, using a data URI in order to provide a thumbnail to the image generated for the NFT.
+     **/
+    function generateImage(
+        uint256 hue,
+        uint256 saturation,
+        uint256 lightness
+    ) public pure returns (string memory) {
         bytes memory svg = abi.encodePacked(
             '<svg width="200" height="300" xmlns="http://www.w3.org/2000/svg">',
-                '<rect width="200" height="300" x="0" y="0" fill="',
-                    'hsl(', 
-                        Strings.toString(hue),
-                        ', ',
-                        Strings.toString(saturation),
-                        '%, ' ,
-                        Strings.toString(lightness),
-                        '%',
-                    ')',
-                '" />',
-            '</svg>'
+            '<rect width="200" height="300" x="0" y="0" fill="',
+            "hsl(",
+            Strings.toString(hue),
+            ", ",
+            Strings.toString(saturation),
+            "%, ",
+            Strings.toString(lightness),
+            "%",
+            ")",
+            '" />',
+            "</svg>"
         );
 
-        return string (
-            abi.encodePacked(
-                'data:image/svg+xml;base64,', 
-                Base64.encode(svg)
-            )
-        );
+        return
+            string(
+                abi.encodePacked(
+                    "data:image/svg+xml;base64,",
+                    Base64.encode(svg)
+                )
+            );
     }
 
     /**
-     * Employs json file delievered by a data URI in order to create the neccesary information about the NFT. 
+     * Employs json file delievered by a data URI in order to create the neccesary information about the NFT.
      **/
-    function generateMetadata (uint nameNum) public pure returns (string memory) {
+    function generateMetadata(uint256 nameNum)
+        public
+        pure
+        returns (string memory)
+    {
         bytes memory json = abi.encodePacked(
-            '{',
+            "{",
             '"name": "ColorCoin #',
             Strings.toString(nameNum),
             '",\n',
-            '"description": "Unique NFTs with color-changing system with every trade. ",\n', 
+            '"description": "Unique NFTs with color-changing system with every trade. ",\n',
             '"image": "',
-            generateImage(6,6,6),
+            generateImage(6, 6, 6),
             '"\n '
-            '}'
-
+            "}"
         );
 
-        return string (
-            abi.encodePacked(
-                'data:application/json;base64,', 
-                Base64.encode(json)
-            )
-        );
+        return
+            string(
+                abi.encodePacked(
+                    "data:application/json;base64,",
+                    Base64.encode(json)
+                )
+            );
     }
+
     /**
-     * This contract creates a random value based on the previous block hash. 
+     * This contract creates a random value based on the previous block hash.
      * DANGEROUS, WILL ONLY GENERATE 1 VALUE PER BLOCK.
      **/
-    function randomNum () public view returns (uint) {
-        return uint(blockhash(block.number - 1));
+    function randomNum() public view returns (uint256) {
+        return uint256(blockhash(block.number - 1));
     }
 
-    function tokenURI(uint tokenId) public pure override returns (string memory) {
-        uint nameNum = tokenId + 1;
+    function tokenURI(uint256 tokenId)
+        public
+        pure
+        override
+        returns (string memory)
+    {
+        uint256 nameNum = tokenId + 1;
         return generateMetadata(nameNum);
     }
-
 }
-
-
