+++ lab/contracts/ColorCoin/ColorCoin.sol
@@ -6,7 +6,7 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 import "@openzeppelin/contracts/utils/Counters.sol";
 
 /**
- * This contract mints 100 NFTs, a list of recipients will recieve the first NFTs, and the rest will go to the creator's wallet. 
+ * This contract mints 100 NFTs, a list of recipients will recieve the first NFTs, and the rest will go to the creator's wallet.
  **/
 contract ColorCoin is ERC721, Ownable {
     using Counters for Counters.Counter;
@@ -14,17 +14,13 @@ contract ColorCoin is ERC721, Ownable {
     Counters.Counter private _tokenIdCounter;
 
     constructor(address[] memory recipients) ERC721("ColorCoin", "CLC") {
-
-        for (uint i = 0; i < 100; i++) {
-
-            if (i < recipients.length){
+        for (uint256 i = 0; i < 100; i++) {
+            if (i < recipients.length) {
                 safeMint(recipients[i]);
-            }
-            else {
+            } else {
                 safeMint(msg.sender);
             }
         }
-
     }
 
     function safeMint(address to) public onlyOwner {
@@ -32,5 +28,4 @@ contract ColorCoin is ERC721, Ownable {
         _tokenIdCounter.increment();
         _safeMint(to, tokenId);
     }
-
 }
