+++ lab/contracts/RandomNum/RandomNum.sol
@@ -2,14 +2,12 @@
 pragma solidity ^0.8.9;
 
 /**
- * This contract creates a random value based on the previous block hash. 
+ * This contract creates a random value based on the previous block hash.
  * DANGEROUS, WILL ONLY GENERATE 1 VALUE PER BLOCK.
  **/
 
 contract RandomNum {
-
-    function randomNum () public view returns (uint) {
-        return uint(blockhash(block.number - 1));
+    function randomNum() public view returns (uint256) {
+        return uint256(blockhash(block.number - 1));
     }
-
 }
